fooda-backend/requirements.txt
django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
pytest
pytest-django
factory_boy
